+++ src/redturtle/volto/restapi/services/controlpanel.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
-from redturtle.volto.interfaces import (
-    IRedTurtleVoltoSettings,
-    IRedTurtleVoltoSettingsControlpanel,
-)
 from plone.restapi.controlpanels import RegistryConfigletPanel
+from redturtle.volto.interfaces import IRedTurtleVoltoSettings
+from redturtle.volto.interfaces import IRedTurtleVoltoSettingsControlpanel
 from zope.component import adapter
-from zope.interface import Interface
 from zope.interface import implementer
+from zope.interface import Interface
 
 
 @adapter(Interface, Interface)