+++ src/redturtle/volto/tests/test_monkeypatches.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
-import unittest
-from Products.CMFPlone.interfaces import ISelectableConstrainTypes
 from plone import api
-from plone.app.testing import (
-    TEST_USER_ID,
-    setRoles,
-)
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from Products.CMFPlone.interfaces import ISelectableConstrainTypes
 from redturtle.volto.testing import REDTURTLE_VOLTO_FUNCTIONAL_TESTING
+
+import unittest
 
 
 class TestRespectLocallyAllowedTypes(unittest.TestCase):