+++ CHANGES.rst
@@ -26,6 +26,12 @@ Changelog
 
 - Plone 6.0 support
 
+- Remove Patch to avoid auto-closed tags in portal transforms: https://github.com/plone/Products.PortalTransforms/pull/43
+  [mamico]
+
+- Remove Patch for folderish types migration https://github.com/plone/plone.volto/pull/86
+  [mamico]
+
 - Handle custom search ranking with AdvancedQuery.
   [mamico]